+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # changelog
 
+* optimize `AbsoluteInt` (with Two's complement implementation)
+* panic if call `AbsoluteInt` with the minimum value of integers (the absolute of the minimum value of integers is impossible to return)
+
 ## v0.3.0
 
 * simplify `ReverseStrings` to reverse original slice